% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{aggregate_candidacies_data}
\alias{aggregate_candidacies_data}
\title{Get candidacies data (at poll station level)}
\source{
...
}
\usage{
aggregate_candidacies_data(
  candidacies_data,
  level = "all",
  id_col_poll = "id_INE_poll_station",
  id_col_candidacies = "id_candidacies",
  id_col_candidacies_prov = "id_candidacies_prov",
  col_abrev_candidacies = "abbrev_candidacies",
  prec_round = 3
)
}
\arguments{
\item{include_candidates}{flag to indicate wheter it should be included data
about candidates or not. Defaults to \code{FALSE}.}
}
\value{
...
}
\description{
...
}
\examples{

## Get candidacies data

# Right examples
\dontrun{
# Wrong examples
}

}
\author{
Javier Álvarez-Liébana.
}
\keyword{get_elections_data}
