% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{aggregate_election_data}
\alias{aggregate_election_data}
\title{Aggregate elections data at provided level (ccaa, prov, etc)}
\usage{
aggregate_election_data(
  election_data,
  level = "all",
  by_parties = TRUE,
  col_id_elec = "id_elec",
  col_id_poll_station = "id_INE_poll_station",
  cols_mun_var = c("pop_res_mun", "census_counting_mun"),
  candidacies_data = NULL,
  col_id_candidacies = "id_candidacies",
  prec_round = 3,
  verbose = TRUE
)
}
\arguments{
\item{election_data}{A database containing general election data
already provided (by other functions or by the user). Database
should contain a \code{col_id_elec} column.
Defaults to \code{NULL}.}

\item{level}{A string providing the level of aggregation at which
the data is to be provided. The allowed values are the following:
'all', 'ccaa', 'prov', 'mun', 'mun_district', 'sec' or
'poll_station'. Defaults to \code{"all"}.}

\item{by_parties}{A flag indicates whether user wants a summary by
candidacies/parties or just global results at given \code{level}.
Defaults to \code{TRUE}.}

\item{col_id_elec}{Column name in \code{election_data} to uniquely
identify the elections. Defaults to \code{"id_elec"}.}

\item{col_id_poll_station}{Column name in \code{ballots_data} to
uniquely identify poll stations. Defaults to
\code{"id_INE_poll_station"}.}

\item{cols_mun_var}{A vector of variable names that, in their raw
version, are only available at the municipal level (or higher).
Defaults to \code{c("pop_res_mun", "census_counting_mun")}.}

\item{candidacies_data}{pending}

\item{col_id_candidacies}{A string indicating the name of the
column that uniquely identifies the candidacies. Defaults to
\code{"id_candidacies"}.}

\item{prec_round}{Rounding accuracy. Defaults to
\code{prec_round = 3}.}

\item{verbose}{Flag to indicate whether detailed messages should
be printed during execution. Defaults to \code{TRUE}.}
}
\value{
A tibble with rows corresponding to the level of aggregation for
each election, including the following variables:
\item{id_elec}{election's id constructed from the election code
\code{cod_elec} and date \code{date_elec}.}
\item{cod_elec}{code representing the type of election:
\code{"01"} (referendum), \code{"02"} (congress),
\code{"03"} (senate), \code{"04"} (local elections),
\code{"06"} (cabildo - Canarian council - elections), \code{"07"}
(European Parliament elections).}
\item{type_elec}{type of election.}
\item{date_elec}{date of the election.}
\item{id_INE_poll_station}{poll station's id constructed from the
ccaa-prov-municipality and poll station codes.}
\item{id_INE_mun}{municipality ID constructed from the
ccaa-prov-mun codes provided by INE.}
\item{cod_INE_ccaa, ccaa}{codes and names for regions (ccaa)
to which the municipalities belong.}
\item{cod_INE_prov, prov}{codes and names for the provinces to
which the municipalities belong.}
\item{cod_INE_mun, mun}{code, and name for
municipalities.}
\item{cod_mun_district, cod_sec, cod_poll_station}{codes for the
municipal district, census tract and poll station.}
\item{ballots_1, turnout_1}{number of total ballots and turnout
percentage in the first round.}
\item{ballots_2, turnout_2}{number of total ballots and turnout
percentage in the second round (if applicable).}
\item{blank_ballots, invalid_ballots}{blank and invalid ballots.}
\item{party_ballots, valid_ballots, total_ballots}{ballots to
candidacies/parties, valid ballots (sum of \code{blank_ballots} and
\code{party_ballots}) and total ballots (sum of
\code{valid_ballots} and \code{invalid_ballots}).}
\item{turnout}{final turnout percentage.}
\item{porc_valid, porc_invalid, porc_parties, porc_blank}{perc (\%)
values of \code{valid_ballots}, \code{invalid_ballots},
\code{party_ballots} and \code{blank_ballots}.}
\item{n_poll_stations}{number of polling stations.}
\item{pop_res_mun}{population census of residents (CER + CERA) at
municipality level.}
\item{census_counting_mun}{population eligible to vote after
claims at municipality level.}
\item{id_candidacies}{id for candidacies (at province level).}
\item{id_candidacies_ccaa, id_candidacies_nat}{id for
candidacies (at region - ccaa - and national level).}
\item{abbrev_candidacies, name_candidacies}{acronym and full name
of the candidacies.}
\item{ballots}{number of ballots obtained for each candidacy at
each level section.}
}
\description{
Aggregate polling station election results
to any chosen territorial level, providing party level ballots,
total ballots, number of polling stations and contextual sums.
}
\details{
...
}
\examples{
## Correct examples

# Election date from 2023 and 1989 dates
election_data <-
   get_election_data(type_elec = "congress", year = c(2023),
                     date = "1989-10-29")

# National level results
nat_agg <- aggregate_election_data(election_data, level = "all")

# National level results (without parties)
nat_agg <-
  aggregate_election_data(election_data, level = "all",
                          by_parties = FALSE)

# Province level results
prov_agg <- aggregate_election_data(election_data, level = "prov")

#  Municipality level results results (without parties)
mun_agg <-
  aggregate_election_data(election_data, level = "mun",
                          by_parties = FALSE)

\dontrun{

# ----
# Incorrect examples
# ----

# Wrong examples

# Invalid 'level' argument,"district" is not allowed

aggregate_election_data(electoral_level, level = "district")

# Invalid 'by_parties' flag: it must be logical, not character
aggregate_election_data(electoral_level, level = "prov",
                        by_parties = "yes")

# Invalid parameters: col_id_candidacies should be matched with
# the variable names
aggregate_election_data(electoral_level, level = "ccaa",
                        col_id_candidacies = "wrong_id")

}

}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{get_elections_data}
