% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_election_data.R
\name{allocate_seats_dhondt}
\alias{allocate_seats_dhondt}
\title{Allocate Seats Using the D'Hondt Method for the Spanish Congress}
\usage{
allocate_seats_dhondt(last_election_ballots, level = "prov")
}
\arguments{
\item{last_election_ballots}{A data frame containing the votes for each polling station.}

\item{level}{The aggregation level for the final seat distribution. Options are \code{"prov"} (default)
for provincial level, \code{"ccaa"} for autonomous community level, and \code{"national"} for national level.}
}
\value{
A data frame with the final seat distribution, filtered to remove candidacies with zero seats.
}
\description{
This function allocates seats to political parties for the Spanish Congress of Deputies
using the D'Hondt method. It allows for aggregation of seat distribution at the provincial,
autonomous community, or national level.
}
\details{
This function performs the following steps:
\enumerate{
\item Aggregates votes by province, autonomous community, and party.
\item Determines the number of seats per province based on the election year using the internal dataset \code{seat_distribution_congress}.
\item Applies the D'Hondt method to allocate seats to parties.
\item Joins the results with party names and autonomous community names from the internal dataset \code{cod_INE_mun}.
\item Aggregates the results based on the specified level.
\item Filters out candidacies with zero seats before returning the final result.
}

#' @author Mikaela DeSmedt
}
\examples{
# Example usage:
final_seat_distribution <- allocate_seats_dhondt(last_election_ballots,
                                                 level = "ccaa")
print(final_seat_distribution)


}
