% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_INE_ccaa}
\alias{cod_INE_ccaa}
\title{INE's code for ccaa and provinces}
\format{
A data frame with 19 rows and 2 variables:
\itemize{
\item \code{cod_INE_ccaa}: code of regions
\item \code{ccaa}: name of regions
}
}
\source{
Data extracted from \href{https://www.ine.es/daco/daco42/codmun/cod_ccaa_provincia.htm}{INE}
}
\usage{
data(cod_INE_ccaa)
}
\description{
A dataset containing the codes provided by INE
of Spanish provinces and regions. Last update: 2023/02/25
}
\examples{
# Load data
data(cod_INE_ccaa)
}
\author{
Data gathered and processed by Javier Álvarez-Liébana.
}
\keyword{datasets}
