% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_INE_mun}
\alias{cod_INE_mun}
\title{Municipal Codes INE}
\format{
A tibble with 8,132 rows and 9 columns:
\describe{
\item{cod_INE_mun}{INE code of the municipality}
\item{mun}{Name of the municipality}
\item{id_INE_mun}{ID in the INE format (e.g., "01-04-001")}
\item{id_MIR_mun}{ID in the MIR format (e.g., "01-04-001")}
\item{cod_INE_prov}{INE code of the province}
\item{prov}{Name of the province}
\item{cod_INE_ccaa}{INE code of the autonomous community (CCAA)}
\item{cod_MIR_ccaa}{MIR code of the autonomous community (CCAA)}
\item{ccaa}{Name of the autonomous community (CCAA)}
}
}
\source{
CCAA $ PROV codes from "https://ine.es/daco/daco42/codmun/cod_provincia.htm", MUN codes from "https://ine.es/daco/daco42/codmun/diccionario24.xlsx"
}
\usage{
data(cod_INE_mun)
}
\description{
Dataset hosting municipality codes for all 8,132 municipalities of spain with their respective CCAA and PROV codes
}
\keyword{INE_codes}
\keyword{datasets}
