% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_INE_prov}
\alias{cod_INE_prov}
\title{INE's code for ccaa and provinces}
\format{
A data frame with 52 rows and 2 variables:
\itemize{
\item \code{cod_INE_prov}: code of province
\item \code{prov}: name of province
}
}
\source{
Data extracted from \href{https://www.ine.es/daco/daco42/codmun/cod_ccaa_provincia.htm}{INE}
}
\usage{
data(cod_INE_prov)
}
\description{
A dataset containing the codes provided by INE
of Spanish provinces and regions. Last update: 2023/02/25
}
\examples{
# Load data
data(cod_INE_prov)
}
\author{
Data gathered and processed by Javier Álvarez-Liébana.
}
\keyword{datasets}
