% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{deans_seats}
\alias{deans_seats}
\title{Function to calculate the allocated seats according to the Dean method in a given
electoral district.}
\usage{
deans_seats(
  candidacies,
  ballots,
  blank_ballots,
  n_seats,
  threshold = 0.03,
  short_version = TRUE
)
}
\arguments{
\item{candidacies}{A vector containing one of the following
variable: unique codes or abbreviations of the candidacies
that participated in the election.}

\item{ballots}{A vector containing the absolute number of ballots
(integer positive values) received by each candidacies}

\item{blank_ballots}{A numerical value indicating the number of
blank ballots (integer positive values).}

\item{n_seats}{An integer positive value indicating the number of
seats that are going to distributed for a given electoral district.}

\item{threshold}{A numerical value (between 0 and 1) indicating the
minimal percentage of votes needed to obtain representation for a
given electoral district. Defaults to \code{0.03}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not. Defaults
to \code{TRUE}.}
}
\value{
A tibble or a list of tibbles  with rows corresponding to each party
including the following variables:
\item{candidacies}{abbrev or id of the candidacies}
\item{seats}{number of seats}
\item{ballots}{absolute number of ballots, just in long format}
\item{porc_seats}{percentage of seats respect to the number of seats,
just in long format}
\item{porc_ballots}{percentage of ballots respect to party ballots
(including blank ballots), just in long format}
\item{quotient_x}{intermediate quotients of the allocation process}
}
\description{
This function allocates seats to political parties in a given electoral district
using the Dean method (also known as the Harmonic Mean method), a highest averages
method for proportional representation. The method uses divisors based on the harmonic mean
of consecutive integers: for each party, its vote total is divided by
\eqn{(2s(s+1))/(2s+1)}, where \eqn{s} is the current number of seats allocated to the party.
Seats are assigned one at a time to the party with the highest resulting quotient
until all seats are distributed. Only parties that exceed a vote threshold
(expressed as a proportion of the total votes, including blank votes) are considered eligible.
}
\details{
The purpose of this helper function is to be used in a general
function, \code{seats_allocation()}, to calculate the seats distribution of every
electoral district of a given election according to the Dean method.
}
\examples{

## Correct examples

## Seats distribution with Dean method for given vectors of candidacies and ballots
## without the remainder quotients

candidacies <- c("PP", "PSOE", "PODEMOS", "VOX")
ballots <- c(200, 350, 100, 200)

seats <- deans_seats(candidacies = candidacies, ballots = ballots,
blank_ballots = 50, n_seats = 15, threshold = 0.03)

\dontrun{

# Incorrect examples

# Different length of candidacies and ballots

candidacies <- c("PP", "PSOE", "PODEMOS", "VOX")
ballots <- c(200, 350, 100)

seats <- deans_seats(candidacies = candidacies, ballots = ballots,
blank_ballots = 50, threshold = 0.03)
}
}
\author{
Javier Alvarez-Liebana, Irene Bosque-Gala and David Pereiro-Pol
}
\keyword{seat_allocation}
