% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{dhondt_seats}
\alias{dhondt_seats}
\title{Function to calculate the allocated seats according to the D'Hont method in a given
electoral district}
\usage{
dhondt_seats(
  parties,
  votes,
  blank_votes,
  nseats,
  threshold,
  short_version = TRUE
)
}
\arguments{
\item{parties}{A vector containing the names or code of the parties that participated in the election.}

\item{votes}{A vector containing the votes received by each party.}

\item{blank_votes}{A vector or number with the blank votes.}

\item{nseats}{A number indicating the number of seats that are going to distributed.}

\item{threshold}{A number indicating the minimal percentage of votes needed to obtain representation.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not.
Defaults to \code{TRUE}.}
}
\value{
A tibble
}
\description{
Calculate seat allocation for a specofoc electoral district following the D'Hont
method that distributes seats proportionally to the votes received by each party.
}
