% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{dhont_seats}
\alias{dhont_seats}
\alias{dhondt_seats}
\title{Function to calculate the allocated seats according to the D'Hont method in a given
electoral district.}
\usage{
dhondt_seats(
  parties,
  votes,
  blank_votes,
  nseats,
  threshold,
  short_version = TRUE
)
}
\arguments{
\item{parties}{A vector containing the names or code of the parties that participated in the election.}

\item{votes}{A vector containing the votes received by each party.}

\item{blank_votes}{A vector or number with the blank votes.}

\item{nseats}{A number indicating the number of seats that are going to distributed.}

\item{threshold}{A number indicating the minimal percentage of votes needed to obtain representation.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not.
Defaults to \code{TRUE}.}
}
\value{
A tibble or a list of tibbles  with rows corresponding to each party including the following
variables:
\item{party}{acronym of the candidacies}
\item{seats}{number of seats distributed to each party}
\item{quotient}{quotients that were not selected for a seat}
}
\description{
This function allocates seats to political parties in a given electoral district
using the D'Hondt method, a highest averages method for proportional representation.
Each party's vote total is divided by a series of divisors (1, 2, 3, ..., up to the number of seats),
generating a list of quotients. Seats are assigned one by one to the highest quotients until all
seats have been distributed. Only parties that surpass a specified vote threshold
(expressed as a proportion of total votes, including blank votes) are eligible for seat allocation.
}
\details{
The purpose of this helper function is to be used in a general
function, \code{seats_allocation()}, to calculate the seats distribution of every
electoral district of a given election according to the D'Hont method.
}
\examples{

## Correct examples

## Seats distribution with D'Hont Method for given vectors of parties and votes
## without the remainder quotients

parties <- c("PP", "PSOE", "PODEMOS", "VOX")
votes <- c(200, 350, 100, 200)

seats <- dhont_seats(parties = parties, votes = votes, blank_votes = 50, threshold = 0.03)

\dontrun

# Incorrect examples

# Different length of parties and votes

parties <- c("PP", "PSOE", "PODEMOS", "VOX")
votes <- c(200, 350, 100)

seats <- dhont_seats(parties = parties, votes = votes, blank_votes = 50, threshold = 0.03)

}
\author{
Javier Alvarez-Liebana, Irene Bosque-Gala and David Pereiro-Pol
}
\keyword{seat_allocation}
