% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{fptp_seats}
\alias{fptp_seats}
\title{Function to calculate the allocated seats according to the First Past
the Post method in a given electoral district.}
\usage{
fptp_seats(
  candidacies,
  ballots,
  blank_ballots,
  n_seats,
  threshold = 0.03,
  short_version = TRUE
)
}
\arguments{
\item{candidacies}{A vector containing one of the following
variable: unique codes or abbreviations of the candidacies
that participated in the election.}

\item{ballots}{A vector containing the absolute number of ballots
(integer positive values) received by each candidacies}

\item{blank_ballots}{A numerical value indicating the number of
blank ballots (integer positive values).}

\item{n_seats}{An integer positive value indicating the number of
seats that are going to distributed for a given electoral district.}

\item{threshold}{A numerical value (between 0 and 1) indicating the
minimal percentage of votes needed to obtain representation for a
given electoral district. Defaults to \code{0.03}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not. Defaults
to \code{TRUE}.}
}
\value{
\item{candidacies}{abbrev or id of the candidacies}
\item{seats}{number of seats distributed to each party}
\item{ballots}{absolute number of ballots, just in long format}
\item{porc_seats}{percentage of seats respect to the number of seats,
just in long format}
\item{porc_ballots}{percentage of ballots respect to party ballots
(including blank ballots), just in long format}
}
\description{
This function allocates every seat in an electoral district to the
candidacy that receives the largest number of ballots. In case of a tie in the number
of votes the seats will be randomnly allocated between those candidacies.
}
\details{
The purpose of this helper function is to be used in a general
function, \code{seats_allocation()}, to calculate the seats distribution of every
electoral district of a given election according to the plurality method.
}
\examples{
## Correct usage ------------------------------------------------------------
candidacies <- c("PP", "PSOE", "PODEMOS", "VOX")
ballots     <- c(200, 350, 100, 200)

fptp_seats(
  candidacies   = candidacies,
  ballots       = ballots,
  blank_ballots = 50,
  n_seats       = 3,
  threshold     = 0.03
)

\dontrun{
## Incorrect input: lengths differ -----------------------------------------
fptp_seats(
  candidacies   = c("PP", "PSOE"),
  ballots       = c(300),
  blank_ballots = 10,
  n_seats       = 1
)
}
}
\author{
Javier Alvarez-Liebana, Irene Bosque-Gala and David Pereiro-Pol
}
\keyword{seat_allocation}
