% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{get_mun_census_data}
\alias{get_mun_census_data}
\title{Get Municipal Census Data}
\source{
Some definitions of variables were extracted from
\url{https://www.ige.gal}.
}
\usage{
get_mun_census_data(type_elec, year, month)
}
\arguments{
\item{type_elec}{A vector or single value representing the types of elections.}

\item{year}{A vector or single value representing the years of the elections to be considered.}

\item{month}{A vector or single value representing the months of the elections to be considered.
Ensure (refer to \code{dates_elections_spain}) that elections of the
specified type are available for the provided year and month.}
}
\value{
A tibble with rows corresponding to municipalities for each election, including the
following columns:
\item{cod_elec}{Code representing the type of election: \code{"01"} (referendum),
\code{"02"} (congress), \code{"03"} (senate), \code{"04"} (local elections),
\code{"06"} (cabildo - Canarian council - elections), \code{"07"}
(European Parliament elections).}
\item{type_elec}{Type of election.}
\item{date_elec}{Date of the election.}
\item{id_INE_mun}{Municipality ID constructed from the ccaa-prov-mun codes provided
by INE.}
\item{id_MIR_mun}{Municipality ID constructed from the ccaa-prov-mun codes provided
by the Spanish Ministry of Interior (MIR).}
\item{cod_INE_ccaa, cod_MIR_ccaa, ccaa}{Codes and names for regions (ccaa)
to which the municipalities belong.}
\item{cod_INE_prov, prov}{Codes and names for the provinces to which the municipalities belong.}
\item{cod_INE_mun, mun}{Code, and name for
municipalities.}
\item{cod_mun_jud_district, cod_mun_prov_council}{Codes for the judicial
district and provincial council.}
\item{n_poll_stations}{Number of polling stations in each municipality.}
\item{pop_res_mun}{Population census of residents (CER + CERA).}
\item{census_INE_mun}{Population from \code{pop_res_mun} eligible to vote.}
\item{census_counting_mun}{Eligible voters after claims.}
\item{census_CERE_mun}{Census of foreign nationals, relevant only for EU elections.}
}
\description{
Fetch municipal census data for one or more elections at the municipal
level. This function downloads and processes raw municipal data files for the specified elections.
}
\details{
This function fetches municipal-level data for the specified elections by downloading
the corresponding \code{.rda} files from GitHub and processing them into a tidy format.
It automatically handles the download, loading, and merging of data across multiple
election periods as specified by the user.
}
\examples{

## Correct examples ----
\dontrun{
# Congress elections in April 2019
# Fetch municipal census data for the congress elections in April 2019
mun_census_data1 <- get_mun_census_data("congress", 2019, 4)

# Senate elections in April 2019
# Fetch municipal census data for the senate elections in April 2019
mun_census_data2 <- get_mun_census_data("senate", 2019, 4)

# Example usage to combine data from different elections into one table
# Fetch municipal census data for congress elections in Nov 2019, April 2019, and June 2016
combined_mun_census_data <- get_mun_census_data(
  c("congress", "congress", "congress"),
  c(2019, 2019, 2016),
  c(11, 4, 6)
)
}
## Incorrect examples ----
\dontrun{
# Invalid election type
# This will fail because "national" is not a valid election type
mun_census_data5 <- get_mun_census_data("national", 2019, 4)

# Length mismatch between year and month vectors
# This will fail because the length of the year and month vectors do not match
mun_census_data6 <- get_mun_census_data("congress", 2016, c(4, 11))

# Invalid date format
# This will fail because the date should be split into year and month
mun_census_data7 <- get_mun_census_data("congress", "2016-06-26")

# Non-existent election data
# This will fail because 1990 congress elections are not available
mun_census_data8 <- get_mun_census_data("congress", 1990, 4)
}
}
\author{
Mikaela DeSmedt, Javier Álvarez-Liébana.
}
\keyword{get_elections_data}
