% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surveys_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get Survey Data from GitHub}
\usage{
get_survey_data(
  from = 1982,
  to = 2023,
  min_days_to = NULL,
  max_days_to = NULL,
  select_polling_firm = "all",
  select_media = "all",
  select_parties = "all",
  min_field_days = NULL,
  max_field_days = NULL,
  min_size = NULL,
  include_media = TRUE,
  include_exit_polls = TRUE
)
}
\arguments{
\item{from}{The start year for fetching surveys (default is 1982). The beginning of the period for which survey data should be retrieved.}

\item{to}{The end year for fetching surveys (default is 2023). The end of the period for which survey data should be retrieved.}

\item{min_days_to}{Minimum number of days to the election for filtering (default is NULL). If specified, only surveys conducted at least this many days before the election are included.}

\item{max_days_to}{Maximum number of days to the election for filtering (default is NULL). If specified, only surveys conducted no more than this many days before the election are included.}

\item{select_polling_firm}{String matching in polling_firm column (default is "all"). If specified, only surveys from polling firms matching this string are included.}

\item{select_media}{String matching in media column (default is "all"). If specified, only surveys from media outlets matching this string are included.}

\item{select_parties}{Character vector specifying which columns (parties) to keep (default is "all"). If specified, only columns corresponding to the specified parties are retained.}

\item{min_field_days}{Minimum number of fieldwork days (default is NULL). If specified, only surveys with at least this many days of fieldwork are included.}

\item{max_field_days}{Maximum number of fieldwork days (default is NULL). If specified, only surveys with no more than this many days of fieldwork are included.}

\item{min_size}{Minimum sample size (default is NULL). If specified, only surveys with at least this sample size are included.}

\item{include_media}{Whether to include media information (default is TRUE). If set to FALSE, media information will be excluded from the results.}

\item{include_exit_polls}{Whether to include exit polls in the data (default is TRUE). If set to FALSE, exit polls will be excluded from the results.}
}
\value{
Data frame of processed survey data. The resulting data frame includes columns such as \code{polling_firm}, \code{media}, \code{sample_size}, \code{turnout}, \code{fieldwork_start}, \code{fieldwork_end}, \code{date_elec}, \code{fieldwork_duration}, \code{is_exit_poll}, \code{party}, and \code{vote_share}.
}
\description{
This function fetches and processes survey data from the specified GitHub directory. It downloads and combines RDA files containing survey data from a specified range of years and months, applies various filters based on the user’s parameters, and returns a cleaned and processed data frame.
}
\details{
The function retrieves and processes survey data files from a specified GitHub directory, applying user-defined filters to the data. The result is a cleaned and combined data frame with relevant survey information, including optional media and exit poll data.
}
\examples{
# Correct usage
survey_data <- get_survey_data(from = 1982, to = 1986, min_days_to = 3)
print(head(survey_data))

# Exclude exit polls
survey_data <- get_survey_data(from = 1982, to = 1986, include_exit_polls = FALSE)
print(head(survey_data))

# Fetching Data from 2000 to 2005 with Maximum Days to Election
survey_data_2 <- get_survey_data(from = 2000, to = 2005, max_days_to = 100)
print(head(survey_data_2))

# Fetching Data from 1990 to 1995 without media
survey_data_3 <- get_survey_data(from = 1990, to = 1995, include_media = FALSE)
print(head(survey_data_3))

# Fetching Data from 1982 to 2020 with Both Minimum and Maximum Days to Election
survey_data_4 <- get_survey_data(from = 1982, to = 2020, min_days_to = 10, max_days_to = 200)
print(head(survey_data_4))

# Fetching Data from 1982 to 2020 with Specific Polling Firm
survey_data_5 <- get_survey_data(from = 1982, to = 2020, select_polling_firm = "CIS")
print(head(survey_data_5))

# Fetching Data from 1982 to 2020 with Specific Media
survey_data_6 <- get_survey_data(from = 1982, to = 2020, select_media = "El País")
print(head(survey_data_6))

# Fetching Data from 1982 to 2020 with Minimum Sample Size
survey_data_7 <- get_survey_data(from = 1982, to = 2020, min_size = 1000)
print(head(survey_data_7))

# Fetching Data from 1982 to 2020 with Specific Parties
survey_data_8 <- get_survey_data(from = 1982, to = 2020, select_parties = c("psoe", "pp"))
print(head(survey_data_8))

# Incorrect usage
# End Year Before Start Year
# This should raise an error or warning
# Reason: The 'to' year is earlier than the 'from' year, which is logically incorrect.
\dontrun{
survey_data_9 <- get_survey_data(from = 1986, to = 1982)
}

# Non-Numeric min_days_to Value
# This should raise an error
# Reason: The min_days_to parameter expects a numeric value, but a string is provided.
\dontrun{
survey_data_10 <- get_survey_data(from = 1982, to = 2023, min_days_to = "three")
}

# Fetching Data Beyond Available Date Range
# This should raise an error or return an empty data frame
# Reason: There are no survey data files available for the specified date range.
\dontrun{
survey_data_11 <- get_survey_data(from = 1900, to = 1910)
}

# Invalid Date Range with Future Dates
# This should raise an error or warning
# Reason: The specified date range includes future years for which no survey data files exist.
\dontrun{
survey_data_12 <- get_survey_data(from = 2025, to = 2030)
}

}
\author{
Mikaela DeSmedt
}
