% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{hagenbach_bischoff_seats}
\alias{hagenbach_bischoff_seats}
\title{Function to calculate the allocated seats according to
the Hagenbach-Bischoff method in a given electoral district.}
\usage{
hagenbach_bischoff_seats(
  candidacies,
  ballots,
  blank_ballots,
  n_seats,
  short_version = TRUE
)
}
\arguments{
\item{candidacies}{A vector containing one of the following
variable: unique codes or abbreviations of the candidacies
that participated in the election.}

\item{ballots}{A vector containing the absolute number of ballots
(integer positive values) received by each candidacies}

\item{blank_ballots}{A numerical value indicating the number of
blank ballots (integer positive values).}

\item{n_seats}{An integer positive value indicating the number of
seats that are going to distributed for a given electoral district.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not. Defaults
to \code{TRUE}.}
}
\value{
A tibble with rows corresponding
to each party including the following variables:
\item{candidacies}{abbrev or id of the candidacies}
\item{seats}{number of seats}
\item{ballots}{absolute number of ballots, just in long format}
\item{porc_seats}{percentage of seats respect to the number of seats,
just in long format}
\item{porc_ballots}{percentage of ballots respect to party ballots
(including blank ballots), just in long format}
\item{remainder}{remainders of the initial division that were
not selected for a seat, just in long format}
}
\description{
This function allocates seats to political parties in a
given electoral district using the Hagenbach-Bischoff method (also
known as the Droop-Hamilton ot simple quota largest-remainders rule).The method first
calculates an electoral quota by dividing the total number
of votes (including blank votes) by the number of seats to be filled plus one.
Each party's vote count is divided by this quota to determine an initial seat
allocation (using the floor of the result). Remaining seats are then assigned
to candidacies with the largest fractional remainders until all seats are distributed.
Only parties that surpass a given vote threshold (expressed as a proportion of
total votes) are considered for seat allocation.
}
\details{
The purpose of this helper function is to be used in a general
function, \code{seats_allocation()}, to calculate the seats distribution of every
electoral district of a given election according to the Hagenbach-Bischoff method.
}
\examples{
## Minimal example ----------------------------------------------------------
candidacies <- c("PP", "PSOE", "PODEMOS", "VOX")
ballots     <- c(200, 350, 100, 200)

hagenbach_bischoff_seats(
  candidacies   = candidacies,
  ballots       = ballots,
  blank_ballots = 50,
  n_seats       = 15
)

\dontrun{
## Incorrect input: vectors of different length ----------------------------
hagenbach_bischoff_seats(
  candidacies   = c("PP", "PSOE", "PODEMOS", "VOX"),
  ballots       = c(200, 350, 100),   # <- length mismatch
  blank_ballots = 50,
  n_seats       = 15
)
}
}
\author{
Javier Alvarez-Liebana, Irene Bosque-Gala and David Pereiro-Pol
}
\keyword{seat_allocation}
