% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{hills_seats}
\alias{hills_seats}
\title{Title}
\usage{
hills_seats(
  parties,
  votes,
  blank_votes,
  nseats,
  threshold,
  short_version = TRUE
)
}
\arguments{
\item{short_version}{}
}
\description{
Title
}
