% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_raw_elections_files}
\alias{import_raw_elections_files}
\alias{import_raw_candidacies_file}
\alias{import_raw_candidates_file}
\alias{import_raw_mun_MIR_files}
\alias{import_poll_stations_MIR_files}
\alias{import_raw_candidacies_poll_file}
\title{Import raw data files of elections}
\source{
Data extracted from
\url{https://infoelectoral.interior.gob.es/opencms/es/elecciones-celebradas/area-de-descargas/}
}
\usage{
import_raw_candidacies_file(
  type_elec,
  year,
  month,
  base_url = "https://infoelectoral.interior.gob.es/estaticos/docxl/apliextr/",
  encoding = "Latin1",
  starts = c(1, 3, 7, 9, 15, 65, 215, 221, 227),
  ends = c(2, 6, 8, 14, 64, 214, 220, 226, 232)
)

import_raw_candidates_file(
  type_elec,
  year,
  month,
  base_url = "https://infoelectoral.interior.gob.es/estaticos/docxl/apliextr/",
  encoding = "Latin1",
  starts = c(1, 3, 7, 9, 10, 12, 13, 16, 22, 25, 26, 51, 76, 101, 102, 104, 106, 110,
    120),
  ends = c(2, 6, 8, 9, 11, 12, 15, 21, 24, 25, 50, 75, 100, 101, 103, 105, 109, 119, 120)
)

import_raw_mun_MIR_files(
  type_elec,
  year,
  month,
  base_url = "https://infoelectoral.interior.gob.es/estaticos/docxl/apliextr/",
  encoding = "Latin1",
  starts = c(1, 3, 7, 10, 12, 14, 17, 19, 120, 123, 129, 137, 142, 150, 158),
  ends = c(2, 6, 8, 11, 13, 16, 18, 118, 122, 125, 136, 141, 149, 157, 165)
)

import_poll_stations_MIR_files(
  type_elec,
  year,
  month,
  base_url = "https://infoelectoral.interior.gob.es/estaticos/docxl/apliextr/",
  encoding = "Latin1",
  starts = c(1, 3, 7, 9, 10, 12, 14, 17, 19, 23, 24, 31, 38, 45, 52, 59, 66, 73, 80),
  ends = c(2, 6, 8, 9, 11, 13, 16, 18, 22, 23, 30, 37, 44, 51, 58, 65, 72, 79, 86)
)

import_raw_candidacies_poll_file(
  type_elec,
  year,
  month,
  base_url = "https://infoelectoral.interior.gob.es/estaticos/docxl/apliextr/",
  encoding = "Latin1",
  starts = c(1, 3, 7, 9, 10, 12, 14, 17, 19, 23, 24, 30),
  ends = c(2, 6, 8, 9, 11, 13, 16, 18, 22, 23, 29, 36)
)
}
\arguments{
\item{type_elec}{type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year, month}{Year and month of election to be unloaded. Please be sure
(see \code{dates_elections_spain}) that elections of the provided type are
available for the given year and month.}

\item{base_url}{A common path from which data will be downloaded}

\item{encoding}{Encoding format to be passed to \code{read_lines()} function}

\item{starts, ends}{Cut-off points for splitting raw .DAT files into a
dataset with several variables, according to the documentation provided by
MIR files.}
}
\value{
A tibble with as many rows as poll stations and a set of the
following variables
\itemize{
\item \code{"01"}: referendum.
\item \code{"02"}: congress.
\item \code{"03"}: senate.
\item \code{"04"}: local elections.
\item \code{"06"}: cabildo - Canarian council - elections).
\item \code{"07"}: European Parlament elections.
}
}
\description{
Import raw data files of elections provided by Spanish Ministry
of Interior (MIR)
}
\examples{
## ...
}
\author{
Javier Álvarez-Liébana.
}
\keyword{utils}
