% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_survey_data.R
\name{import_survey_data}
\alias{import_survey_data}
\title{Import survey data}
\usage{
import_survey_data(
  type_elec = "congress",
  year = NULL,
  date = NULL,
  verbose = TRUE
)
}
\arguments{
\item{type_elec}{Type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year}{A vector or single value representing the years of the
elections to be considered. Please, check in
\code{dates_elections_spain} that elections of the specified type
are available for the provided year.}

\item{date}{A vector or single value representing the dates of
the elections to be considered. If date was provided, it should be
in format \%Y-\%m-\%d (e.g., '2000-01-01'). Defaults to \code{NULL}.
If no date was provided, \code{year} should be provided as
numerical variable. Please, check in \code{dates_elections_spain}
that elections of the specified type are available.}

\item{verbose}{Flag to indicate whether detailed messages should
be printed during execution. Defaults to \code{TRUE}.}
}
\value{
A tibble with rows corresponding to the estimated results
for each party for each election given by a particular pollster,
including the following variables:

\item{id_survey}{survey's id constructed from the polling
firm and the dates for the start and end of the fieldwork.}
\item{id_elec}{election's id constructed from the election code
and date.}
\item{polling_firm}{organisation conducting the poll.}
\item{media}{commissioning organisation / media outlet.}
\item{fieldwork_start, fieldwork_end, n_field_days}{fieldwork period
and its length (days).}
\item{sample_size}{sample size of the survey.}
\item{abbrev_candidacies, name_candidacies_nat}{acronym and full name
of the candidacies.}
\item{id_candidacies_nat}{id for candidacies at national level.}
\item{estimated_porc_ballots}{estimated percentage of ballots
for each party.}
}
\description{
Import and preprocess surveys data for given election
types and dates. This function supports
both single values and vector inputs for fetching and combining
data for multiple elections' dates at once.
}
\details{
This function fetches survey data for the
specified elections' date by loading the corresponding files from
\code{{pollspaindata}} and processing them by joining the ids and name
of the different political parties.
}
\examples{

## Correct examples

# Fetch survey data for congress elections in multiple dates
poll_station_data <-
  import_survey_data(year = 2023,
                           date = "2019-04-28")

# ----
# Incorrect examples
# ----

\dontrun{
# Wrong examples

# Invalid election type: "national" is not a valid election type
import_survey_data(type_elec = "national", year = 2019)

# Invalid election: no congress elections are available in 2018
# Please check dataset dates_elections_spain
import_survey_data(type_elec = "congress", year = 2018)

# Invalid date format: date should be in \%Y-\%m-\%d format
import_survey_data(tdate = "26-06-2016")

}

}
\author{
Javier Alvarez-Liebana and David Pereiro Pol
}
\keyword{import_survey_data}
