% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{party_colors_hex}
\alias{party_colors_hex}
\title{Party Colors Hex Data}
\format{
A tibble with 21 rows and 3 columns:
\describe{
\item{abbrev_candidacies}{Abbreviation of the candidacy}
\item{name_candidacies}{Full name of the candidacy}
\item{party_color}{Hex color code associated with the party}
}
}
\source{
Manually sourced by Mikaela DeSmedt
}
\usage{
data(party_colors_hex)
}
\description{
Dataset holding hexadecimal codes for spanish political parties
}
\keyword{colors}
\keyword{datasets}
\keyword{party}
