% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_election_data.R
\name{plot_election_results}
\alias{plot_election_results}
\title{Plot Spanish Congress Election Results}
\usage{
plot_election_results(
  election_data,
  level = "prov",
  colors_url =
    "https://github.com/mikadsr/Pollspain-data/raw/main/get\%20auxiliary\%20data/party_colors_hex.rda"
)
}
\arguments{
\item{election_data}{A data frame containing the election results. It should include
the following columns:
\describe{
\item{cod_INE_prov}{Character. The INE code for provinces.}
\item{prov}{Character. The name of the province.}
\item{cod_MIR_ccaa}{Character. The MIR code for autonomous communities.}
\item{abbrev_candidacies}{Character. The abbreviated name of the political party.}
\item{name_candidacies}{Character. The full name of the political party.}
\item{ccaa}{Character. The name of the autonomous community.}
\item{ballots}{Numeric. The number of ballots received by the party.}
}}

\item{level}{A character string indicating the geographic level to plot.
Options are \code{"prov"} (default) for province-level results, and \code{"ccaa"} for
autonomous community-level results.}

\item{colors_url}{A character string specifying the URL from which to
download the color codes for the political parties. Default is
"https://github.com/mikadsr/Pollspain-data/raw/main/get\%20auxiliary\%20data/party_colors_hex.rda".}
}
\value{
A ggplot2 object showing the election results on a map of Spain.
}
\description{
This function generates a map of Spain, visualizing the election results
for the Spanish Congress of Deputies. The map can display results by province
or autonomous community (CCAA) based on the \code{level} argument. The function
colors each region according to the party with the most votes in that area.
}
\details{
This function performs the following steps:
\enumerate{
\item Downloads the color codes for political parties from the specified URL.
\item Processes the election data to match the required format for plotting.
\item Generates a map of Spain with regions colored according to the party with the most votes.
\item Customizes the map based on the specified \code{level} parameter.
}
}
\examples{
\dontrun{
# Example usage with election data
plot_election_results(election_data = a2023, level = "prov")
plot_election_results(election_data = a2023, level = "ccaa")
}


}
\author{
Mikaela DeSmedt
}
