% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_election_data.R
\name{plot_parliament_distribution}
\alias{plot_parliament_distribution}
\title{Plot Parliamentary Seat Distribution}
\usage{
plot_parliament_distribution(
  election_data,
  colors_url =
    "https://github.com/mikadsr/Pollspain-data/raw/main/get\%20auxiliary\%20data/party_colors_hex.rda"
)
}
\arguments{
\item{election_data}{A data frame containing election results with the following structure:
\itemize{
\item \code{cod_INE_prov}: Province code
\item \code{prov}: Province name
\item \code{cod_MIR_ccaa}: Autonomous community code
\item \code{abbrev_candidacies}: Party abbreviations
\item \code{name_candidacies}: Full party names
\item \code{ccaa}: Autonomous community name
\item \code{seats}: Number of seats won by each party
}}

\item{colors_url}{A string URL pointing to the location of the RDA file containing party colors. Default is set to "https://github.com/mikadsr/Pollspain-data/raw/main/get\%20auxiliary\%20data/party_colors_hex.rda".}
}
\value{
A ggplot object representing the parliamentary seat distribution.
}
\description{
This function generates a semicircle plot of parliamentary seat distribution for a given election dataset using the \code{ggparliament} package. It also fetches custom party colors from an external source to enhance the visual representation of the results.
}
\details{
This function performs the following steps:
\enumerate{
\item Downloads the party color information from the specified URL.
\item Processes the election data to prepare it for visualization.
\item Generates a semicircle plot showing the distribution of parliamentary seats among parties.
\item Customizes the plot using the downloaded party colors and visualizes the seat allocation effectively.
}
}
\examples{
\dontrun{
# Load example data
data(a)
# Plot the seat distribution
plot_parliament_distribution(election_data = a)
}


}
\author{
Mikaela DeSmedt, Javier Álvarez-Liébana
}
