% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollspain-package.R
\docType{package}
\name{pollspain-package}
\alias{pollspain-package}
\alias{pollspain}
\title{\pkg{pollspain} -- A grammar for the analysis of electoral
and poll data}
\description{
The package is intended to provide a tidyverse-style
grammar for preprocessing, debugging, filtering, accessing,
downloading and analyzing election data and polls in Spain.
}
\references{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dadosdelaplace/pollspain}
  \item \url{http://javieralvarezliebana.es/pollspain/}
  \item Report bugs at \url{https://github.com/dadosdelaplace/pollspain/issues}
}

}
\author{
Javier Álvarez-Liébana, David Pereiro-Pol,
Mafalda González-González, Irene Bosque-Gala and Mikaela De Smedt.
}
