% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recod_parties}
\alias{recod_parties}
\title{Recode Political Party Names and Abbreviations}
\usage{
recod_parties(
  parties_data,
  col_name_abbrev = "abbrev_candidacies",
  col_name_candidacies = "name_candidacies"
)
}
\arguments{
\item{parties_data}{A dataframe containing political party data with columns for party abbreviations and full names.}

\item{col_name_abbrev}{A character string specifying the column name for party abbreviations in the input dataframe. Defaults to "abbrev_candidacies".}

\item{col_name_candidacies}{A character string specifying the column name for party full names in the input dataframe. Defaults to "name_candidacies".}
}
\value{
A dataframe with standardized and recoded political party names and abbreviations. The output dataframe includes cleaned versions of the original columns and recoded values based on predefined patterns.
}
\description{
This function standardizes and recodes the names and abbreviations of political parties in a given dataset. It handles various text transformations, such as removing special characters, trimming whitespace, and converting characters to ASCII. The function also recodes certain party names and abbreviations to standardized formats based on predefined patterns.
}
\details{
The function first renames the specified columns in the input dataframe to a standard format ("abbrev_candidacies" and "name_candidacies"). It then performs a series of text cleaning operations, such as removing special characters, replacing certain symbols with hyphens, and converting accented characters to their ASCII equivalents. After cleaning the text, the function applies a series of pattern-matching rules to recode specific party names and abbreviations to standardized versions.
}
\examples{
\dontrun{
# Assuming `party_data` is a dataframe with columns "abbrev" and "full_name"
standardized_data <- recod_parties(party_data, col_name_abbrev = "abbrev", col_name_candidacies = "full_name")
}

}
\author{
De Smedt, Mikaela; Alvarez-Liebana, Javier
}
