% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{seat_allocation}
\alias{seat_allocation}
\title{Function to compute the allocated seats according to the
chosen method for a given electoral districts.}
\usage{
seat_allocation(
  candidacies,
  ballots,
  blank_ballots,
  n_seats,
  method = "hondt",
  threshold = 0.03,
  short_version = TRUE
)
}
\arguments{
\item{candidacies}{A vector containing one of the following
variable: unique codes or abbreviations of the candidacies
that participated in the election.}

\item{ballots}{A vector containing the absolute number of ballots
(integer positive values) received by each candidacies}

\item{blank_ballots}{A numerical value indicating the number of
blank ballots (integer positive values).}

\item{n_seats}{An integer positive value indicating the number of
seats that are going to distributed for a given electoral district.}

\item{method}{A string vector providing the methods of
apportionment to be used. The allowed values are the following:
\code{"D'Hondt"} (or \code{"Hondt"} or \code{"hondt"}),
\code{"Hamilton"} (or \code{"hamilton"} or \code{"Vinton"} or
\code{"vinton"}), \code{"Webster"} (or \code{"webster"} or
\code{"Sainte-Lague"} or \code{"sainte-lague"}), \code{"Hill"} (or
\code{"hill"} or \code{"Huntington-Hill"} or
\code{"huntington-hill"}), \code{"Dean"} (or \code{"dean"}) or
\code{"Adams"} (or \code{"adams"}) or
\code{"Hagenbach-Bischoff"} (or \code{"hagenbach"})
(or \code{"bischoff"}) or
\code{"First Past the Post"} (or \code{"first"}) (or \code{"fptp"}).
Defaults to \code{"Hondt"}.}

\item{threshold}{A numerical value (between 0 and 1) indicating the
minimal percentage of votes needed to obtain representation for a
given electoral district. Defaults to \code{0.03}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not. Defaults
to \code{TRUE}.}
}
\value{
A tibble with rows corresponding to each party including
the following variables:
\item{candidacies}{abbrev or id of the candidacies}
\item{seats}{number of seats}
\item{method}{method to allocate seats}
\item{ballots}{absolute number of ballots, just in long format}
\item{porc_ballots}{percentage of ballots respect to party ballots
(including blank ballots), just in long format}
\item{quotient_x}{intermediate quotients of the allocation process}
}
\description{
Function to compute the allocated seats according to the
chosen method for a given electoral districts.
}
\details{
The purpose of this function is to calculate the
allocation of seats for a given electoral district (and given a
particular election) for a set of apportionment methods.
}
\examples{

## Correct examples

# Seats distribution with Hill method for given vectors of
# candidacies and ballots without the remainder quotients
candidacies <- c("PP", "PSOE", "PODEMOS", "VOX")
ballots <- c(200, 350, 100, 200)
seats <- seat_allocation(method = "hill", candidacies = candidacies,
                        ballots = ballots, blank_ballots = 50,
                        n_seats = 15, threshold = 0.03)

# Same results in a long version (providing quotients)
seats <- seat_allocation(method = "hill", candidacies = candidacies,
                        ballots = ballots, blank_ballots = 50,
                        n_seats = 15, threshold = 0.03, short_version = FALSE)

# Adams with threshold 0.05
seats <- seat_allocation(method = "adams",candidacies, ballots,
                        blank_ballots = 50, n_seats = 5, threshold = 0.05)

# A very high threshold that only one party meets in Dean method
seats <- seat_allocation(method = "dean",candidacies, ballots,
                        blank_ballots = 50, n_seats = 5, threshold = 0.3)

\dontrun{

# ----
# Incorrect examples
# ----

# Different length of candidacies and ballots

candidacies <- c("PP", "PSOE", "PODEMOS", "VOX")
ballots <- c(200, 350, 100)

seats <- seat_allocation(method = "hill", candidacies = candidacies,
ballots = ballots, blank_ballots = 50, threshold = 0.03)

 # Ballots with missing values
candidacies <- c("PP", "PSOE", "PODEMOS")
ballots <- c(200, 350, NA)
seats <- seat_allocation(method = "hagenbach", candidacies = candidacies,
ballots = ballots, blank_ballots = 50)

# Ballots with char values
candidacies <- c("PP", "PSOE", "PODEMOS")
ballots <- c("200", "350", "100")
seats <- seat_allocation(method = "hamilton", candidacies = candidacies,
ballots = ballots, blank_ballots = 50)

# Threshold should be a numerical value between 0 and 1
candidacies <- c("PP", "PSOE", "PODEMOS")
ballots <- c(200, 350, 100)
seats <- seat_allocation(method = "webster", candidacies = candidacies,
ballots = ballots, blank_ballots = 50, threshold = 3)

}
}
\author{
David Pereiro-Pol, Irene Bosque-Gala and Javier
Alvarez-Liebana.
}
\keyword{seat_allocation}
