% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seat_distribution_congress}
\alias{seat_distribution_congress}
\title{Seat Distribution Congress Data}
\format{
A data frame with 676 rows and 4 columns:
\describe{
\item{prov}{Province name}
\item{cod_INE_prov}{INE code of the province}
\item{year}{Year of the election}
\item{seats}{Number of seats allocated}
}
}
\source{
Mannualy sourced by Mikaela DeSmedt from "https://es.wikipedia.org/wiki/Circunscripciones_electorales_del_Congreso_de_los_Diputados"
}
\usage{
data(seat_distribution_congress)
}
\description{
Dataset holding the number of seats distributed by each constituency from 1982 until 2023.
}
\keyword{congress}
\keyword{datasets}
\keyword{seat-distribution}
