% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat_allocation.R
\name{seats}
\alias{seats}
\title{Function to compute the allocated seats according to the
chosen method for a given electoral districts.}
\usage{
seats(
  candidacies,
  ballots,
  blank_ballots,
  n_seats,
  method = "hondt",
  threshold = 0.03,
  short_version = TRUE
)
}
\arguments{
\item{candidacies}{A vector containing one of the following
variable: unique codes or abbreviations of the candidacies
that participated in the election.}

\item{ballots}{A vector containing the absolute number of ballots
(integer positive values) received by each candidacies}

\item{blank_ballots}{A numerical value indicating the number of
blank ballots (integer positive values).}

\item{n_seats}{An integer positive value indicating the number of
seats that are going to distributed for a given electoral district.}

\item{method}{A string vector providing the methods of
apportionment to be used. The allowed values are the following:
\code{"D'Hondt"} (or \code{"Hondt"} or \code{"hondt"}),
\code{"Hamilton"} (or \code{"hamilton"} or \code{"Vinton"} or
\code{"vinton"}), \code{"Webster"} (or \code{"webster"} or
\code{"Sainte-Lague"} or \code{"sainte-lague"}), \code{"Hill"} (or
\code{"hill"} or \code{"Huntington-Hill"} or
\code{"huntington-hill"}), \code{"Dean"} (or \code{"dean"}) and
\code{"Adams"} (or \code{"adams"}) . Defaults to \code{"Hondt"}.}

\item{threshold}{A numerical value (between 0 and 1) indicating the
minimal percentage of votes needed to obtain representation for a
given electoral district. Defaults to \code{0.03}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not. Defaults
to \code{TRUE}.}
}
\value{
A tibble with rows corresponding to each party including
the following variables:
\item{candidacies}{abbrev or id of the candidacies}
\item{seats}{number of seats}
\item{method}{method to allocate seats}
\item{ballots}{absolute number of ballots, just in long format}
\item{porc_ballots}{percentage of ballots respect to party ballots
(including blank ballots), just in long format}
\item{quotient_x}{intermediate quotients of the allocation process}
}
\description{
Function to compute the allocated seats according to the
chosen method for a given electoral districts.
}
\details{
The purpose of this function is to calculate the
allocation of seats for a given electoral district (and given a
particular election) for a set of apportionment methods.
}
\examples{

##Correct examples

\dontrun{

# ----
# Incorrect examples
# ----

}
}
\author{
Irene Bosque-Gala, David Pereiro-Pol and Javier
Alvarez-Liebana.
}
\keyword{seats}
