% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seats_by_province}
\alias{seats_by_province}
\title{Number of parliament seats allocated to each part for 10 provinces
for the general Spanish elections.}
\format{
A tibble with 11 columns
\itemize{
\item \code{id_elec}: id of elections (type of elections + date).
\item \code{cod_INE_ccaa, cod_MIR_ccaa, ccaa}: codes and names
for ccaa.
\item \code{cod_INE_prov, prov}: codes and names for provinces.
\item \code{id_INE_prov}: id for prov (\code{cod_INE_ccaa} +
\code{cod_INE_prov})
\item \code{abbrev_candidacies}: party abbreviation.
\item \code{id_candidacies_nat}: national id.
\item \code{id_candidacies}: id at each constituency.
\item \code{real_seats}: number of seats allocated to each party.
}
}
\source{
Data extracted from
\url{https://infoelectoral.interior.gob.es/es/inicio/}
}
\usage{
data(seats_by_province)
}
\description{
A dataset containing the parliament seats for 10
electoral districts (province) by party and election, just considering
congress seats.
}
\examples{
# Load data
data(seats_by_province)
}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{datasets}
