% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_survey_data.R
\name{summary_survey_data}
\alias{summary_survey_data}
\title{Summaries of the poll data for
a given election.}
\usage{
summary_survey_data(
  type_elec = "congress",
  year = NULL,
  date = NULL,
  verbose = TRUE,
  short_version = TRUE
)
}
\arguments{
\item{type_elec}{Type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year}{A vector or single value representing the years of the
elections to be considered. Please, check in
\code{dates_elections_spain} that elections of the specified type
are available for the provided year.}

\item{date}{A vector or single value representing the dates of
the elections to be considered. If date was provided, it should be
in format \%Y-\%m-\%d (e.g., '2000-01-01'). Defaults to \code{NULL}.
If no date was provided, \code{year} should be provided as
numerical variable. Please, check in \code{dates_elections_spain}
that elections of the specified type are available.}

\item{verbose}{Flag to indicate whether detailed messages should
be printed during execution. Defaults to \code{TRUE}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not.
Defaults to \code{TRUE}.}
}
\value{
A tibble with rows corresponding to the estimated results
for each party for each election given by a particular pollster,
including the following variables:

\item{id_survey}{survey's id constructed from the polling
firm and the dates for the start and end of the fieldwork.}
\item{id_elec}{election's id constructed from the election code
and date.}
\item{polling_firm}{organisation conducting the poll.}
\item{media}{commissioning organisation / media outlet.}
\item{n_polls}{polls done by each pollster for a particular election}
\item{fieldwork_start, fieldwork_end, n_field_days}{fieldwork period
and its length (days).}
\item{days_until_election}{number of days until the next election
counting form the start of the fieldwork}
\item{sample_size}{sample size of the survey.}
\item{abbrev_candidacies, name_candidacies_nat}{acronym and full name
of the candidacies at national level.}
\item{id_candidacies_nat}{id for candidacies at national level.}
\item{estimated_porc_ballots}{estimated percentage of ballots
for each party.}
}
\description{
Import, preprocess and summary survey data.
}
\details{
This function uses the helper function \code{import_survey_data()},
which imports and cleans survey data, and adds other relevant variables for
analysis. This function gives a tidy table of the survey data that
is ready for analysis or visualisation.
}
\examples{

## Correct examples

# Summary 2023 survey data at prov level in short version

summary_survey_data(year = 2023,
                        short_version = TRUE)

\dontrun{
# ----
# Incorrect examples
# ----

# Wrong examples

# Invalid year
summary_survey_data(year = 2018, short_version = FALSE)

# Invalid short version flag: short_version should be a
# logical variable
summary_survey_data(type_elec = "congress", year = 2019,
                  short_version = "yes")
}

}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{summary_survey_data}
