% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_seats_spain}
\alias{total_seats_spain}
\title{Number of Parliament seats allocated to each province for
the general elections.}
\format{
A tibble with 4 columns
\itemize{
\item \code{id_elec}: id of elections (type of elections + date).
\item \code{id_INE_prov, prov}: codes and names
for provinces.
\item \code{seats}: number of seats allocated to each province.
}
}
\source{
pending...
}
\usage{
data(total_seats_spain)
}
\description{
A dataset containing the Parliament seats per
electoral district (province)and election. It is used to calculate
seat allocation with the standard distribution of seats in Spain.
}
\examples{
# Load data
data(total_seats_spain)
}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{datasets}
