% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{type_to_code_election}
\alias{type_to_code_election}
\title{Conversion type election to code}
\source{
Data extracted from
\url{https://infoelectoral.interior.gob.es/opencms/es/elecciones-celebradas/area-de-descargas/}{Spanish Ministry of Interior}
}
\usage{
type_to_code_election(type_elec)
}
\arguments{
\item{type_elec}{type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}
}
\value{
A vector of string codes (\code{cod_elec}) as follows:
\itemize{
\item \code{"01"}: referendum.
\item \code{"02"}: congress.
\item \code{"03"}: senate.
\item \code{"04"}: local elections.
\item \code{"06"}: cabildo - Canarian council - elections).
\item \code{"07"}: European Parlament elections.
}
}
\description{
Conversion the type of election (referendum,
congress, etc) to a properly code (according to documentation)
available in Spanish Ministry of Interior
}
\examples{
## Convert type to code

# Right examples
type_to_code_election(type_elec = "congress")
type_to_code_election(type_elec = "senate")
type_to_code_election(type_elec = "local")

\dontrun{
# Wrong examples
type_to_code_election(type_elec = "hi!")
type_to_code_election(type_elec = "nop")
type_to_code_election(type_elec = "national")
}

}
\author{
Mikaela DeSmedt;Javier Alvarez-Liebana
}
\keyword{utils}
