% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{aggregate_election_data}
\alias{aggregate_election_data}
\title{Get candidacies data (at poll station level)}
\source{
...
}
\usage{
aggregate_election_data(
  election_data,
  level = "all",
  id_col = "id_INE_poll_station",
  cod_CERA = "999",
  prec_round = 3
)
}
\arguments{
\item{include_candidates}{flag to indicate wheter it should be included data
about candidates or not. Defaults to \code{FALSE}.}
}
\value{
...
}
\description{
...
}
\examples{

## Get candidacies data

# Right examples
\dontrun{
# Wrong examples
}

}
\author{
Javier Álvarez-Liébana.
}
\keyword{get_elections_data}
