% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{aggregate_election_data}
\alias{aggregate_election_data}
\title{Aggregate elections data}
\source{
...
}
\usage{
aggregate_election_data(
  election_data,
  level = "all",
  by_parties = TRUE,
  prec_round = 3
)
}
\arguments{
\item{election_data}{tibble containing election data}

\item{level}{level of territorial aggregation. Values allowed are "all"
(national aggregation), "ccaa" (regional aggregation),
"prov" (province aggregation), "mun" (municipal aggregation),
"mun-district" (municipal district aggregation),
"sec" (census section aggregation) or "poll-station" (without aggregation)}

\item{by_parties}{flag to indicate wheter it should be aggregate data
about ballots of parties (as long as \code{election_data} has been generated
with the candidacies data). Defaults to \code{TRUE}.}

\item{prec_round}{rounding accuracy. Defaults to \code{prec_round = 3}.}
}
\value{
...
}
\description{
...
}
\examples{

## Aggregate election data

# Right examples
\dontrun{
# Wrong examples
}

}
\author{
Javier Álvarez-Liébana.
}
\keyword{get_elections_data}
