% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_INE_mun}
\alias{cod_INE_mun}
\title{INE's code for municipalities}
\format{
A tibble with 8131 rows and 10 variables:
\itemize{
\item \code{id_INE_ccaa, id_MIR_ccaa}: full id of municipalities
(combining ccaa, province and mun's id)
\item \code{cod_INE_ccaa, cod_MIR_ccaa}: code of regions
\item \code{ccaa}: name of regions
\item \code{cod_INE_prov}: code of provinces
\item \code{prov}: name of provinces
\item \code{cod_INE_mun}: code of municipalities
\item \code{cd_INE_mun}: check digit (see \href{https://www.ine.es/daco/daco42/codmun/codmun00i.htm}{documentation})
\item \code{mun}: name of municipalities
}
}
\source{
Data extracted from \href{https://www.ine.es/daco/daco42/codmun/codmun20/20codmun.xlsx}{INE}
}
\usage{
data(cod_INE_mun)
}
\description{
A dataset containing the codes provided by INE
of Spanish municipalities. Last update: 2025/05/03 (data
updated by INE on 2020/01/01)
}
\examples{
# Load data
data(cod_INE_mun)
}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{datasets}
