% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod_INE_prov_ccaa}
\alias{cod_INE_prov_ccaa}
\title{INE's code for ccaa and provinces}
\format{
A tibble with 52 rows and 5 variables:
\itemize{
\item \code{cod_INE_ccaa}: code of region according INE.
\item \code{cod_MIR_ccaa}: code of region according Spanish
Ministry of the Interior (MIR).
\item \code{ccaa}: name of region
\item \code{cod_INE_prov}: code of province
\item \code{prov}: name of province
}
}
\source{
Data extracted from
\href{https://www.ine.es/daco/daco42/codmun/cod_ccaa_provincia.htm}{INE}
}
\usage{
data(cod_INE_prov_ccaa)
}
\description{
A dataset containing the codes provided by INE
of Spanish provinces and regions. Last update: 2023/02/25
}
\examples{
# Load data
data(cod_INE_prov_ccaa)
}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{datasets}
