% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_years}
\alias{detect_years}
\title{Detect election date}
\usage{
detect_years(type_elec = "congress", year = NULL, date = NULL)
}
\arguments{
\item{type_elec}{Type of the election. Default \code{type_elec = "congress"}.}

\item{year}{A vector or single value representing the years of the
elections to be considered. Please, check in
\code{dates_elections_spain} that elections of the specified type
are available for the provided year.}

\item{date}{}
}
\description{
Given a year and/or date, this function
returns the a data frame with information on the day of the elections.
}
\details{
This function is used as a helper function to select the dates
asked for the users in the import and summary functions.
}
\author{
David Pereiro Pol.
}
\keyword{utils}
