% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{get_elections_data}
\alias{get_elections_data}
\title{Get candidacies data (at poll station level)}
\source{
...
}
\usage{
get_elections_data(
  type_elec,
  year,
  month,
  level = "all",
  by_parties = TRUE,
  include_candidacies = FALSE,
  include_candidates = FALSE,
  id_col = "id_INE_poll_station",
  prec_round = 3
)
}
\arguments{
\item{type_elec}{type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year, month}{vector of years and months of elections to be considered.
Please be sure (see \code{dates_elections_spain}) that elections of the
provided type are available for the given year and month.}

\item{include_candidates}{flag to indicate wheter it should be included data
about candidates or not. Defaults to \code{FALSE}.}
}
\value{
...
}
\description{
...
}
\examples{

## Get candidacies data

# Right examples
\dontrun{
# Wrong examples
}

}
\author{
Javier Álvarez-Liébana.
}
\keyword{get_elections_data}
