% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{get_poll_station_data}
\alias{get_poll_station_data}
\title{Get poll station data}
\source{
...
}
\usage{
get_poll_station_data(type_elec, year, month, prec_round = 3)
}
\arguments{
\item{type_elec}{type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year, month}{vector of years and months of elections to be considered.
Please be sure (see \code{dates_elections_spain}) that elections of the
provided type are available for the given year and month.}

\item{prec_round}{rounding accuracy. Defaults to \code{prec_round = 3}.}
}
\value{
...
}
\description{
...
}
\examples{

## Get poll station data

# Right examples
\dontrun{
# Wrong examples
}

}
\author{
Javier Álvarez-Liébana.
}
\keyword{get_elections_data}
