% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surveys_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get Survey Data from Pollspain-data}
\usage{
get_survey_data(
  year,
  min_days_to = NULL,
  max_days_to = NULL,
  select_polling_firm = "all",
  select_media = "all",
  select_parties = "all",
  min_field_days = NULL,
  max_field_days = NULL,
  min_size = NULL,
  include_media = TRUE,
  include_exit_polls = TRUE
)
}
\arguments{
\item{year}{A single year or a vector of years for which survey data should be retrieved.}

\item{min_days_to}{Minimum number of days to the election for filtering (default is NULL). If specified, only surveys conducted at least this many days before the election are included.}

\item{max_days_to}{Maximum number of days to the election for filtering (default is NULL). If specified, only surveys conducted no more than this many days before the election are included.}

\item{select_polling_firm}{String matching in polling_firm column (default is "all"). If specified, only surveys from polling firms matching this string are included.}

\item{select_media}{String matching in media column (default is "all"). If specified, only surveys from media outlets matching this string are included.}

\item{select_parties}{Character vector specifying which parties to keep (default is "all"). If specified, only columns corresponding to the specified parties are retained.}

\item{min_field_days}{Minimum number of fieldwork days (default is NULL). If specified, only surveys with at least this many days of fieldwork are included.}

\item{max_field_days}{Maximum number of fieldwork days (default is NULL). If specified, only surveys with no more than this many days of fieldwork are included.}

\item{min_size}{Minimum sample size (default is NULL). If specified, only surveys with at least this sample size are included.}

\item{include_media}{Whether to include media information (default is TRUE). If set to FALSE, media information will be excluded from the results.}

\item{include_exit_polls}{Whether to include exit polls in the data (default is TRUE). If set to FALSE, exit polls will be excluded from the results.}
}
\value{
A data frame of processed survey data. The resulting data frame includes columns such as \code{polling_firm}, \code{media}, \code{sample_size}, \code{turnout}, \code{fieldwork_start}, \code{fieldwork_end}, \code{date_elec}, \code{fieldwork_duration}, \code{is_exit_poll}, \code{party}, and \code{vote_share}.
}
\description{
This function fetches and processes survey data from the Pollspain-data repository It downloads and combines RDA files containing survey data from a specified year or range of years, applies various filters based on the user’s parameters, and returns a cleaned and processed data frame.
}
\details{
The function retrieves and processes survey data files from a specified GitHub directory, applying user-defined filters to the data. The result is a cleaned and combined data frame with relevant survey information, including optional media and exit poll data. The function now accepts a \code{year} argument that can be a single year or a vector of years, replacing the previous \code{from} and \code{to} parameters. It verifies the availability of \code{congress} elections in the specified years using the \code{dates_elections_spain} dataset.
}
\examples{
# Fetch survey data for a specific year
survey_data <- get_survey_data(year = 1982, min_days_to = 3)
print(head(survey_data))

# Exclude exit polls for multiple years
survey_data <- get_survey_data(year = c(1982, 1986), include_exit_polls = FALSE)
print(head(survey_data))

# Fetching Data for a range of years with Maximum Days to Election
survey_data_2 <- get_survey_data(year = 2000:2005, max_days_to = 100)
print(head(survey_data_2))

# Fetching Data for a range of years without media
survey_data_3 <- get_survey_data(year = 1990:1995, include_media = FALSE)
print(head(survey_data_3))

# Fetching Data for specific years with Specific Polling Firm
survey_data_4 <- get_survey_data(year = c(1982, 1983), select_polling_firm = "CIS")
print(head(survey_data_4))

# Fetching Data with Minimum Sample Size
survey_data_5 <- get_survey_data(year = 1982:2020, min_size = 1000)
print(head(survey_data_5))

# Fetching Data with Specific Parties
survey_data_6 <- get_survey_data(year = 1982:2020, select_parties = c("psoe", "pp"))
print(head(survey_data_6))

# Incorrect usage
# Fetching Data for a year where no congress election exists
# This should raise an error or return an empty data frame
\dontrun{
survey_data_7 <- get_survey_data(year = 1900)
}

# Invalid year range
# This should raise an error or warning
# Reason: The specified years include future dates for which no survey data files exist.
\dontrun{
survey_data_8 <- get_survey_data(year = 2025:2030)
}

}
\author{
Mikaela DeSmedt
}
