% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_dict_parties}
\alias{global_dict_parties}
\title{pending}
\format{
A data frame with pending:
\itemize{
\item \code{cod_INE_ccaa}: code of regions
\item \code{ccaa}: name of regions
\item \code{cod_INE_prov}: code of provinces
\item \code{prov}: name of provinces
\item \code{cod_INE_mun}: code of municipalities
\item \code{cd_INE_mun}: check digit (see \href{https://www.ine.es/daco/daco42/codmun/codmun00i.htm}{documentation})
\item \code{mun}: name of municipalities
}
}
\source{
Data ...
}
\usage{
data(global_dict_parties)
}
\description{
pending
}
\examples{
# Load data
data(global_dict_parties)
}
\author{
Irene Bosque-Gala, Mafalda Gonzalez-Gonzalez and Javier
Alvarez-Liebana.
}
\keyword{datasets}
