% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_elections_data.R
\name{import_mun_census_data}
\alias{import_mun_census_data}
\title{Import municipal census data}
\source{
Some definitions of variables were extracted from
\url{https://www.ige.gal}.
}
\usage{
import_mun_census_data(
  type_elec,
  year = NULL,
  date = NULL,
  verbose = TRUE,
  lazy_duckdb = FALSE
)
}
\arguments{
\item{type_elec}{Type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year}{A vector or single value representing the years of the
elections to be considered. Please, check in
\code{dates_elections_spain} that elections of the specified type
are available for the provided year.}

\item{date}{A vector or single value representing the dates of
the elections to be considered. If date was provided, it should be
in format \%Y-\%m-\%d (e.g., '2000-01-01'). Defaults to \code{NULL}.
If no date was provided, \code{year} should be provided as
numerical variable. Please, check in \code{dates_elections_spain}
that elections of the specified type are available.}

\item{verbose}{Flag to indicate whether detailed messages should
be printed during execution. Defaults to \code{TRUE}.}

\item{lazy_duckdb}{Flag to indicate whether a lazy duckDB database
should be provided (\code{TRUE}) or a tibble (\code{FALSE}).
Defaults to \code{FALSE}.}
}
\value{
A tibble with rows corresponding to municipalities for
each election, including the following variables:
\item{id_elec}{election's id constructed from the election code
\code{cod_elec} and date \code{date_elec}.}
\item{cod_elec}{code representing the type of election: \code{"01"}
(referendum), \code{"02"} (congress), \code{"03"} (senate),
\code{"04"} (local elections), \code{"06"} (cabildo - Canarian
council - elections), \code{"07"} (European Parliament elections).}
\item{type_elec}{type of election.}
\item{date_elec}{date of the election.}
\item{id_INE_mun}{municipality ID constructed from the
ccaa-prov-mun codes provided by INE.}
\item{cod_INE_ccaa, ccaa}{codes and names for regions (ccaa)
to which the municipalities belong.}
\item{cod_INE_prov, prov}{codes and names for the provinces to
which the municipalities belong.}
\item{cod_INE_mun, mun}{code, and name for municipalities.}
\item{cod_mun_jud_district, cod_mun_prov_council}{codes for the
judicial district and provincial council.}
\item{n_poll_stations}{number of polling stations in each
municipality.}
\item{pop_res_mun}{population census of residents (CER + CERA)
at municipality level.}
\item{census_INE_mun}{population eligible to vote at municipality
level.}
\item{census_counting_mun}{population eligible to vote after
claims at municipality level.}
\item{census_CERE_mun}{census of foreign nationals, relevant only
for EU elections at municipality level.}
}
\description{
Import municipal census data for one or more elections
at the municipal level (from 1982 onwards). This function
downloads and processes raw municipal data files for specified
elections. No CERA data is included as municipalities (note that
CERA ballots will be included in the summaries (see
\code{import_poll_station_data()} and
\code{import_candidacies_data()})
}
\details{
This function fetches municipal-level data for the
specified elections by downloading the corresponding files from
\code{{pollspaindata}} package and processing them into a tidy format.
It automatically handles the download, loading, and merging
of data across multiple election periods as specified by the user.
}
\examples{

## Correct examples

# Fetch municipal census for congress elections in multiple dates
mun_census <-
  import_mun_census_data(type_elec = "congress",
                         year = c(2016, 2023), date = "2019-04-28")

# ----
# Incorrect examples
# ----
\dontrun{
# Wrong examples

# Invalid election type: "national" is not a valid election type
import_mun_census_data("national", 2019)

# Invalid election: no congress elections are available in 5-2019.
# Please check dataset dates_elections_spain
import_mun_census_data("congress", date = "2019-05-01")

#' # Invalid election: no congress elections are available in 2018.
# Please check dataset dates_elections_spain
import_mun_census_data("congress", year = 2018)

# Invalid date format: date should be in \%Y-\%m-\%d format
import_mun_census_data("congress", date = "26-06-2016")
}
}
\author{
Javier Alvarez-Liebana, David Pereiro Pol, Mafalda Gonzalez
Gonzalez, Irene Bosque Gala and Mikaela De Smedt.
}
\keyword{import_elections_data}
