% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollspain-package.R
\docType{"_PACKAGE"}
\name{pollspain-package}
\alias{pollspain-package}
\alias{pollspain}
\title{pollspain}
\description{
An R package to access, clean, and analyze electoral and survey data for the Spanish elections from 1982 until present.
}
\details{
This package includes functionality to work with electoral data, survey data, and various auxiliary datasets related to Spanish elections.
}
