% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollspain-package.R
\docType{package}
\name{pollspain-package}
\alias{pollspain-package}
\alias{pollspain}
\title{\pkg{pollspain} -- blabla}
\description{
bla bla
}
\references{
García-Portugués, E., Álvarez-Liébana, J., Álvarez-Pérez, G. and
González-Manteiga, W. (2019). A goodness-of-fit test for the functional
linear model with functional response. \emph{arXiv:1909.07686}.
\url{https://arxiv.org/abs/1909.07686}

García-Portugués, E., González-Manteiga, W. and Febrero-Bande, M. (2014). A
goodness-of-fit test for the functional linear model with scalar response.
\emph{Journal of Computational and Graphical Statistics}, 23(3):761--778.
\url{http://doi.org/10.1080/10618600.2013.812519}
}
\author{
Javier Álvarez-Liébana.
}
