% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recod_parties}
\alias{recod_parties}
\title{Recode of party's names}
\usage{
recod_parties(
  parties_data,
  col_name_abbrev = "abbrev_candidacies",
  col_name_candidacies = "name_candidacies"
)
}
\arguments{
\item{parties_data}{...}

\item{col_name_abbrev}{...}

\item{col_name_candidacies}{...}
}
\value{
...
}
\description{
...
}
\author{
Javier Álvarez-Liébana.
}
\keyword{utils}
