% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recod_parties}
\alias{recod_parties}
\title{Recoding Spanish political party or candidacies names and
preprocessing of acronyms}
\arguments{
\item{parties_data}{A dataset (preferably a tibble) containing
political party data. Function will expect a dataset with 2 columns
for party acronyms and full names.}

\item{col_acronym}{A character string specifying the column name
for party acronyms in \code{parties_data}. Defaults to "abbrev_candidacies".}

\item{col_full_name}{A character string specifying the column name
for party full names in \code{parties_data}. Defaults to "name_candidacies".}
}
\value{
A tibble with standardized and recoded political party
names and acronyms. The output datasets includes cleaned versions
of the original columns and recoded values based on predefined
patterns.
}
\description{
This function standardizes and recodes the names and
abbreviations of Spanish political party names in a given dataset.
It handles various text transformations, such as removing special
characters, trimming whitespace, and converting characters to ASCII.
The function also recodes certain party names and abbreviations to
standardized formats based on predefined patterns.
}
\details{
The function first renames the specified columns in the
input dataset to a standard format. It then performs a series of
text cleaning operations, such as removing special characters,
replacing certain symbols with hyphens, and converting accented
characters to their ASCII equivalents. After cleaning the text,
the function applies a series of pattern-matching rules to recode
specific party names and abbreviations to standardized versions.
}
\examples{
\dontrun{
# Assuming `party_data` is a dataframe with columns
# "abbrev" and "full_name"
standardized_data <-
  recod_parties(party_data, col_name_abbrev = "abbrev",
                col_name_candidacies = "full_name")
}

}
\author{
Javier Álvarez-Liébana, David Pereiro-Pol, Mafalda
González-González and Irene Bosque-Gala.
}
\keyword{utils}
