% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elections_data.R
\name{summary_election_data}
\alias{summary_election_data}
\title{Summaries of the electoral and candidacies ballots data for
a given aggregation level (ccaa, prov, etc)}
\usage{
summary_election_data(
  type_elec,
  year = 2019,
  date = NULL,
  election_data = NULL,
  ballots_data = NULL,
  col_id_elec = "id_elec",
  col_id_poll_station = "id_INE_poll_station",
  prec_round = 3,
  short_version = TRUE,
  repo_url = "https://github.com/dadosdelaplace/pollspain-data/blob/main/data",
  file_ext = ".rda",
  level = "all",
  by_parties = FALSE,
  filter_porc_ballots = NA,
  cols_mun_var = c("pop_res_mun", "census_counting_mun"),
  col_id_candidacies = "id_candidacies_nat",
  cols_names_candidacies = c("abbrev_candidacies", "name_candidacies"),
  verbose = TRUE
)
}
\arguments{
\item{type_elec}{Type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year}{A vector or single value representing the years of the
elections to be considered. Please, check in
\code{dates_elections_spain} that elections of the specified type
are available for the provided year.}

\item{date}{A vector or single value representing the dates of
the elections to be considered. If date was provided, it should be
in format \%Y-\%m-\%d (e.g., '2000-01-01'). Defaults to \code{NULL}.
If no date was provided, \code{year} should be provided as
numerical variable. Please, check in \code{dates_elections_spain}
that elections of the specified type are available.}

\item{election_data}{A database containing general election data
already provided (by other functions or by the user). Database
should contain a \code{col_id_elec} column.
Defaults to \code{NULL}.}

\item{ballots_data}{A database containing ballots data for
the different candidacies, already provided (by other functions or
by the user). Database should contain a
\code{col_id_poll_station} column. Defaults to \code{NULL}.}

\item{col_id_elec}{Column name in \code{election_data} to uniquely
identify the elections. Defaults to \code{"id_elec"}.}

\item{col_id_poll_station}{Column name in \code{ballots_data} to
uniquely identify poll stations. Defaults to
\code{"id_INE_poll_station"}.}

\item{prec_round}{Rounding accuracy. Defaults to
\code{prec_round = 3}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not.
Defaults to \code{TRUE}.}

\item{repo_url}{A string with the url in which the raw data can be
found. Defaults to
\url{https://github.com/dadosdelaplace/pollspain-data/blob/main}.}

\item{file_ext}{A string with the file's extension. Defaults to
\code{".rda"}.}

\item{level}{A string providing the level of aggregation at which
the data is to be provided. The allowe values are the following:
'all', 'ccaa', 'prov', 'mun', 'mun_district', 'sec' or
'poll_station'. Defaults to \code{"all"}.}

\item{by_parties}{A flag indicates whether user wants a summary by
candidacies/parties or just global results at given \code{level}.
Defaults to \code{FALSE}.}

\item{filter_porc_ballots}{A numerical argument representing the
vote percentage threshold (out of 100) that the user wants to use
to filter the parties (as long as \code{by_parties = TRUE}).
Defaults to \code{NA}.}

\item{cols_mun_var}{A vector of variable names that, in their raw
version, are only available at the municipal level (or higher).
Defaults to \code{c("pop_res_mun", "census_counting_mun")}.}

\item{col_id_candidacies}{A string indicating the name of the
column that uniquely identifies the candidacies. Defaults to
\code{"id_candidacies"}.}

\item{cols_names_candidacies}{A string indicating the name of the
column containg the names and/or acronyms of candidacies.
Defaults to \code{c("abbrev_candidacies", "name_candidacies")}.}

\item{verbose}{Flag to indicate whether detailed messages should
be printed during execution. Defaults to \code{TRUE}.}
}
\value{
A tibble with rows corresponding to ...pending
}
\description{
pending...
}
\details{
This function ... pending
}
\examples{

## Correct examples

# Summary election data at national level (general data
# without candidacies ballots)
summary_data_all <- summary_election_data("congress", 2019)

# Summary election data at national level, aggregating the
# candidacies ballots
summary_data_all_parties <-
  summary_election_data("congress", 2019, 4, by_parties = TRUE)

# Summary election data at ccaa level (general data
# without candidacies ballots)
summary_data_ccaa <-
  summary_election_data("congress", 2019, 4, level = "ccaa")

# Summary election data at ccaa level, aggregating the
# candidacies ballots
summary_data_ccaa_parties <-
  summary_election_data("congress", 2019, 4, level = "ccaa",
                        by_parties = TRUE)

\dontrun{
# ----
# Incorrect examples
# ----

# Wrong examples

}

}
\author{
Javier Álvarez-Liébana and David Pereiro-Pol.
}
\keyword{get_elections_data}
