% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{summary_survey_data}
\alias{summary_survey_data}
\title{Summaries of the poll data for a given election.}
\usage{
summary_survey_data(
  type_elec = "congress",
  year = NULL,
  date = NULL,
  verbose = TRUE,
  short_version = TRUE,
  format = "long",
  forthcoming = FALSE,
  filter_polling_firm = NULL,
  filter_media = NULL,
  filter_n_field_days = NULL,
  filter_days_until_elec = NULL,
  lower_sample_size = NULL,
  upper_sample_size = NULL,
  lower_fieldwork_date = NULL,
  upper_fieldwork_date = NULL,
  filter_abbrev_candidacies = NULL
)
}
\arguments{
\item{type_elec}{Type elections for which data is available.
It should be one of the following values: "referendum",
"congress", "senate", "local", "cabildo" (Canarian council)
or "EU".}

\item{year}{A vector or single value representing the years of the
elections to be considered. Please, check in
\code{dates_elections_spain} that elections of the specified type
are available for the provided year.}

\item{date}{A vector or single value representing the dates of
the elections to be considered. If date was provided, it should be
in format \%Y-\%m-\%d (e.g., '2000-01-01'). Defaults to \code{NULL}.
If no date was provided, \code{year} should be provided as
numerical variable. Please, check in \code{dates_elections_spain}
that elections of the specified type are available.}

\item{verbose}{Flag to indicate whether detailed messages should
be printed during execution. Defaults to \code{TRUE}.}

\item{short_version}{Flag to indicate whether it should be returned
a short version of the data (just key variables) or not.
Defaults to \code{TRUE}.}

\item{format}{Do you want the output in long or wide format?
Defaults to \code{"long"}}

\item{forthcoming}{A flag indicates whether user wants to include
surveys for the forthcoming elections. Defaults to \code{TRUE}. If
\code{TRUE}, no date neither year are required (in that case, just
surveys for the next elections are provided).}

\item{filter_polling_firm, filter_media, filter_abbrev_candidacies}{Do you want to filter surveys by polling firm, media or abbrev of
candidacies? A string vector should be introduced. Defaults to
\code{NULL}.}

\item{filter_n_field_days, filter_days_until_elec}{Do you want to
filter surveys by number of fieldwork days or days until election?
A single numeric value should be introduced. Defaults to \code{NULL}.}

\item{lower_sample_size, upper_sample_size}{Do you want to
filter surveys by sample size? Single numeric values should be
introduced for each one. Defaults to \code{NULL} in both cases.}

\item{lower_fieldwork_date, upper_fieldwork_date}{Do you want to
filter surveys by fieldwork_date? Single date values should be
introduced for each one. Defaults to \code{NULL} in both cases.}
}
\value{
A tibble with rows corresponding to the estimated results
for each party for each election given by a particular pollster,
including the following variables:

\item{id_survey}{survey's id constructed from the polling
firm and the dates for the start and end of the fieldwork.}
\item{id_elec}{election's id constructed from the election code
and date.}
\item{polling_firm}{organisation conducting the poll.}
\item{media}{commissioning organisation / media outlet. Variable
not available for short version.}
\item{n_polls_by_elec}{Number of polls done by each pollster for a
particular election. Variable not available for short version.}
\item{fieldwork_start, fieldwork_end, n_field_days}{fieldwork period
and its length (days).}
\item{days_until_election}{number of days until the next election
counting form the start of the fieldwork. Variable not available
for short version.}
\item{sample_size}{sample size of the survey.}
\item{abbrev_candidacies, name_candidacies_nat}{acronym and full name
of the candidacies at national level. The last one is not available
available for short version.}
\item{id_candidacies_nat}{id for candidacies at national level.}
\item{estimated_porc_ballots}{estimated percentage of ballots
for each party.}
}
\description{
Import, preprocess and summary survey data.
}
\details{
This function uses the helper function \code{import_survey_data()},
which imports and cleans survey data, and adds other relevant variables for
analysis. This function gives a tidy table of the survey data that
is ready for analysis or visualisation.
}
\examples{

## Correct examples

# Summary for all 2019-2023 surveys
summary_survey_data(year = c(2019, 2023))

# Summary for 2019-2023 surveys in a full version, filtering
# just 40DB and GAD3 polling firms, the last 30 days before elec,
# and with a sample size at least than 500 people
summary_survey_data(year = c(2019, 2023), short_version = FALSE,
                    filter_polling_firm = c("GAD3", "40DB"),
                    filter_days_until_elec = 30,
                    lower_sample_size =  500)

# Summary for surveys for the forthcoming elections in, filtering
# by parties ("PP" and "PSOE") and with a sample size of at least
# 500 people
summary_survey_data(forthcoming = TRUE,
                    filter_abbrev_candidacies = c("PP", "PSOE"),
                    lower_sample_size =  500)

# Summary for 2023 surveys in a full version, filtering
# just 40DB and GAD3 polling firms, surveys that go from the
# first of March of 2020 to the first of August of the same year
summary_survey_data(year = 2023, short_version = FALSE,
                    filter_polling_firm = c("GAD3", "40DB"),
                    lower_fieldwork_date = "2020-03-01",
                    upper_fieldwork_date = "2020-08-01")
\dontrun{
# ----
# Incorrect examples
# ----

# Wrong examples

# Invalid year
summary_survey_data(year = 2018, short_version = FALSE)

# Invalid short version flag: short_version should be a
# logical variable
summary_survey_data(type_elec = "congress", year = 2019,
                  short_version = "yes")

# upper_fieldworkday prior to lower_fieldworkday
summary_survey_data(year = 2023,
                    lower_fieldwork_day = "2020-08-01",
                    upper_fieldwork_day = "2020-03-01")

# upper_sample_size smaller than lower_sample-size
summary_survey_data(year = 2023,
                    lower_fieldwork_day = 500,
                    upper_fieldwork_day = 200)

# Invalid class for argument filter_polling_firm
summary_survey_data(year = 2023,
                    filter_polling_firm = 700)

# Invalid class for argument filter_media
summary_survey_data(year = 2023,
                    filter_media = 700)


}
}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{summary_survey_data}
