% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_seats_spain}
\alias{total_seats_spain}
\title{Number of parliament seats allocated to each province for
the general Spanish elections.}
\format{
A tibble with 8 columns
\itemize{
\item \code{id_elec}: id of elections (type of elections + date).
\item \code{cod_INE_ccaa, cod_MIR_ccaa, ccaa}: codes and names
for ccaa.
\item \code{cod_INE_prov, prov}: codes and names for provinces.
\item \code{id_INE_prov}: id for prov (\code{cod_INE_ccaa} +
\code{cod_INE_prov})
\item \code{nseats}: number of seats allocated to each province.
}
}
\source{
Data extracted from
\url{https://www.juntaelectoralcentral.es/cs/jec/elecciones}
}
\usage{
data(total_seats_spain)
}
\description{
A dataset containing the parliament seats for each
electoral district (province) and election, just considering
congress seats.
}
\examples{
# Load data
data(total_seats_spain)
}
\author{
Javier Alvarez-Liebana and David Pereiro-Pol.
}
\keyword{datasets}
